+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Copy (Ctrl+C) action options.
 - Alternative Copy (Ctrl+Shift+C) action options.
 - Alternative Copy Fallback (Ctrl+Shift+C with no selection) action options.
+- Trim option for all copy actions.
 - 
 
 ### Changed